fix(ui): refactor and translations
