+++ ui/src/lib/i18n/locales/en.json
@@ -93,6 +93,8 @@
   "No Parameters Available": "No Parameters Available",
   "Current Version": "",
   "Version ": "",
+  "Are you absolutely sure you want to change version?": "",
+  "Test description here": "",
   "Unable to rename workflow": "Unable to rename workflow",
   "Renaming workflow failed. Please try again later.": "Renaming workflow failed. Please try again later.",
   "Main Workflow": "",
