+++ ui/src/lib/i18n/locales/es.json
@@ -93,6 +93,8 @@
   "No Parameters Available": "No hay parámetros disponibles",
   "Current Version": "",
   "Version ": "Versión ",
+  "Are you absolutely sure you want to change version?": "",
+  "Test description here": "",
   "Unable to rename workflow": "No se puede renombrar el flujo de trabajo",
   "Renaming workflow failed. Please try again later.": "Renombrar el flujo de trabajo falló. Por favor, inténtalo de nuevo más tarde.",
   "Main Workflow": "Flujo de trabajo principal",
