+++ ui/src/lib/i18n/locales/fr.json
@@ -93,6 +93,8 @@
   "No Parameters Available": "Aucun paramètre disponible",
   "Current Version": "",
   "Version ": "Version ",
+  "Are you absolutely sure you want to change version?": "",
+  "Test description here": "",
   "Unable to rename workflow": "Impossible de renommer le flux de travail",
   "Renaming workflow failed. Please try again later.": "Échec du renommage du flux de travail. Veuillez réessayer plus tard.",
   "Main Workflow": "Flux de travail principal",
