+++ ui/src/lib/i18n/locales/ja.json
@@ -93,6 +93,8 @@
   "No Parameters Available": "利用可能なパラメータはありません",
   "Current Version": "現在のバージョン",
   "Version ": "バージョン ",
+  "Are you absolutely sure you want to change version?": "",
+  "Test description here": "",
   "Unable to rename workflow": "ワークフローの名前を変更できません",
   "Renaming workflow failed. Please try again later.": "ワークフローの名前変更に失敗しました。後でもう一度お試しください。",
   "Main Workflow": "メインワークフロー",
