+++ ui/src/lib/i18n/locales/zh.json
@@ -93,6 +93,8 @@
   "No Parameters Available": "没有可用参数",
   "Current Version": "",
   "Version ": "版本 ",
+  "Are you absolutely sure you want to change version?": "",
+  "Test description here": "",
   "Unable to rename workflow": "无法重命名工作流",
   "Renaming workflow failed. Please try again later.": "重命名工作流失败。请稍后再试。",
   "Main Workflow": "主工作流",
